refactor: mobile drawer menu imports, interface & index exports
